Update resource building mechanics
Add setup.py commands for building resources: "build_resources" and
"build_production_resources".

Make the "build" command call "build_production_resources" as a
subcommand.  This makes "setup.py bdist_wheel", "pip wheel",
"pip install", and other similar commands to build the resources.

The building is done with build_resources.py, which replaces the old
build-all.py.  Most notable difference is the caching that is implented
here, because we don't want to build the same resources over and over
again in the test server.

Other smaller changes included to achieve this:

 * Update MANIFEST.in so that source files and scripts to build resources
   are included to sdist.
 * Add generated_resources.txt to shoop/admin and shoop/notify so that
   build_resources.py knows what to cache.
 * Tune EXCLUDE_PATTERNS in setup.py.
 * Define walk_excl and is_excluded_filename in setup.py to be used by
   build_resources.py.

Refs SHOOP-994