tox.ini: Add changedir setting back
The changedir setting was removed in commit d7a4a6a97bf, because some
Tox version between 1.0 and 2.3 did not support it and I thought it is
not needed.  Actually we need the changedir after all:  Without it the
coverage will not be collected correctly and the destination dir for
coverage.xml in project root is problematic, because it should be
different for each Tox environment.

Add the changedir back and use {toxinidir}/shoop_tests to point py.test
to correct path of the tests.  The coverage.xml will now be generated to
{envsitepackagesdir} again.

Refs SHOOP-1954