Ensure pricing context gets shop and customer
While getting rid of OrderSource.request and updating tests for it, some
tests failed, because they passed in None as customer, but the error was
hard to track down.

Make it more explicit that shop and customer are required by adding
couple asserts to PricingContext and making them required parameters for
`PricingModule.get_context_from_data`.