Admin: Fix contact detail page with many groups
If contact belongs to multiple groups, the contact details views used to
fail on Python 3, since there was a call `contact.groups.all()|sort` in
the template, which tries to sort ContactGroup objects.  Fix this by
adding sorted contacts to the context and remove sorting from the
template.

Also while at it:

  * Render the contact groups as links.
  * Add test case for this.
  * Fix ContactGroupQuerySet to derive from TranslatableQuerySet.

Refs SHOOP-2397