Fix contact group availability behavior component
All customers belong to the "Anonymous" group by default (AKA. not signed in users),
thus it's pointless to check for the existance of the user in the database. This
also causes limiting to not logged in users only to be impossible.

Fix the issue explained above by removing a check for customer's existance in the database.