Add allow-root to the bower config in the Dockerfile
By default, bower disallows running it as root. On a traditional
system this makes sense: it uses the unix user model to separate
bower packages from the system. By not being root, the scripts can't
really attack the system. In Docker, the separation is provided by the
container system, so running foreign scripts as root isn't as much of
an issue. By being in a container, foreign scripts can't attack the
system.

Allowing root disables the user-based separation model.