Core: Order Status improvements
Order statuses were used by Shuup in a way where
merchant could not modify them at all. This commit will enable the
modification and adding of the Order Statuses.

This commit also adds `OrderStatusManager` which is a helper class
that ensures that default order statuses are always available. Future
plans with this class is that it handles all the order status changes
with plug-in status change rules.

Shuup now has 4 default order statuses:
* Initial
* Processing
* Completed
* Canceled

These order statuses cannot be modified (only name and public
name can be changed) and they must always be present in shuup.

Breakdown:
* Make `OrderStatus` objects manageable
* Add `OrderStatusManager` to handle and identify default order statuses
* Add `is_active` and `public_name` fields to `OrderStatus`
* Add Admin view for order statuses
* Modify `shuup_init` to always update order statuses
* Add `is_partially_shipped` and `is_partially_paid` methods to order

Refs SH-432