xtheme: Fix Social Media Links plugin
The SoMe links plugin couldn't be saved because it tried to access the
translated fields incorrectly from the cleaned_data dictionary in its
clean method.

This for-loop

    for field in plugin_fields:
        cleaned_data[field] = precleaned_data.pop(field)

tried to access, e.g. field named "topic" even though there was no such
field in the precleaned_data dictionary.  There was only "topic_*",
"topic_en", "topic_fi", etc.

Fix this by rewriting the clean method to not touch the fields it
doesn't understand, but rather juts rewrite the links, which it knows.