front tests: Update phase view initialization
Amend the tests which used to create a phase view directly to create the
phase view through a checkout view so that the checkout_process is
passed to the constructor of the view.

We want to avoid constructing the phase views without a checkout
process, since that is deprecated.  The tests currently pass before and
after this change though.