Provide user activation link for the company approved event template
Provide user activation link for the company approved event template
so it can be used to send "Company has been approved" message
containing activation link for the user to activate his account.
This way admin only needs to activate the CompanyContact and it will
trigger the CompanyApproved event which sends the activation link
for the user to activate his own account.

Create a new signal which will be dispatched when the company is
activated through the admin panel. This signal has the request
object which is used to build the activation url. Connect the
send_company_activated_first_time_notification function to this
signal instead of post_save on CompanyContact.