Hotfix: fix cache signal handler when called through m2m
In some scenario shop product post save signal handler
is called with category. Handle this situation and
bump all products in that category.

There is round of improvements reserved for this context
cache bumping and goal is to put all the logic in one
place which is not in any signal handler. Also we look
into making this cache bumping as effortles as possible.