Fix `Uncaught ReferenceError: sync is not defined` error
This was initially caught by using the `Product Selection` plugin,
because the existing customer wondered how to add entries (the
`Products` field was empty and not selectable; despite there being a lot
of products added in the back-end already) in that plugin and if some
kind of add/select button existed as he remembered.

Browser consoles gave `Uncaught ReferenceError: sync is not defined`
error.

The long story short: the error was fixed by removing reference to
`sync` that was used but wasn't used anywhere.

This tiny change also fixed a bunch of other issues we didn't even know
we had with Select2 widget.

This commit also simultaneously fixes:
- all other plugins that use Select2 widget like Category Products
Highlight, Product Selection, Category Links, CMS Page Links, etc.
- the plugin selection choice (now you can also type the name of the
plugin, instead of using the dropmenu)
- all other selection fields in the front-end.

refs SHU-45