Fix bug: page jumps after selecting tabs
If the tab's content is small (fitting on one screen), then clicking on 
that tab caused the view jump to page's top. This is good.

But if the tab didn't fit on one screen, then clicking on the tab cause 
a view to jump to that tab section hashtag anchor link location 
(url/#something). This sounds good, but in practice it was inconsistent 
and caused a graphical bug (important content not being seen).

refs SHU-36