Front: ensure company information on checkout
Ensure company name and tax number is set to both billing and shipping
address same way as when filled through company form when customer
is not logged in. Company name and tax number at order addresses seems
to help with some taxation logic as well as makes things more
consistent.